+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.11.0] - 2026-06-04 - Kitty Terminal graphics protocol support
+
 ### Added
 
 - `moops.run` accepts `output_mode=` to control where a child notebook's